Remove walletkit
