+++ .github/CODEOWNERS
@@ -11,7 +11,6 @@
 /packages/notification-service/ @annakaz @cmcewen
 /packages/protocol/ @asaj @m-chrzan
 /packages/react-components/ @cmcewen @jmrossy
-/packages/walletkit/ @ashishb @yerdua
 /packages/terraform-modules/ @tkporter
 /packages/transaction-metrics-exporter @nambrot
 /packages/typescript/ @cmcewen
