+++ packages/celotool/package.json
@@ -8,7 +8,6 @@
   "dependencies": {
     "@celo/verification-pool-api": "^1.0.0",
     "@celo/utils": "0.1.9-dev",
-    "@celo/walletkit": "^0.0.15",
     "@celo/contractkit": "0.3.2-dev",
     "@google-cloud/monitoring": "0.7.1",
     "@google-cloud/pubsub": "^0.28.1",
