+++ packages/celotool/src/lib/geth.ts
@@ -6,7 +6,6 @@ import { TransactionResult } from '@celo/contractkit/lib/utils/tx-result'
 import { GoldTokenWrapper } from '@celo/contractkit/lib/wrappers/GoldTokenWrapper'
 import { StableTokenWrapper } from '@celo/contractkit/lib/wrappers/StableTokenWrapper'
 import { waitForPortOpen } from '@celo/dev-utils/lib/network'
-import { unlockAccount } from '@celo/walletkit'
 import BigNumber from 'bignumber.js'
 import { spawn } from 'child_process'
 import fs from 'fs'
@@ -14,6 +13,7 @@ import { merge, range } from 'lodash'
 import fetch from 'node-fetch'
 import path from 'path'
 import sleep from 'sleep-promise'
+import Web3 from 'web3'
 import { TransactionReceipt } from 'web3-core'
 import { Admin } from 'web3-eth-admin'
 import { convertToContractDecimals } from './contract-utils'
@@ -30,6 +30,20 @@ import { GethInstanceConfig } from './interfaces/geth-instance-config'
 import { GethRunConfig } from './interfaces/geth-run-config'
 import { ensure0x, spawnCmd, spawnCmdWithExitOnFailure } from './utils'
 import { getTestnetOutputs } from './vm-testnet-utils'
+
+export async function unlockAccount(
+  web3: Web3,
+  duration: number,
+  password: string,
+  accountAddress: string | null = null
+) {
+  if (accountAddress === null) {
+    const accounts = await web3.eth.getAccounts()
+    accountAddress = accounts[0]
+  }
+  await web3.eth.personal.unlockAccount(accountAddress!, password, duration)
+  return accountAddress!
+}
 
 type HandleErrorCallback = (isError: boolean, data: { location: string; error: string }) => void
