+++ packages/mobile/metro.config.js
@@ -8,7 +8,7 @@ const cwd = path.resolve(__dirname)
 const root = path.resolve(cwd, '../..')
 const escapedRoot = escapeStringRegexp(root)
 const celoRegex = new RegExp(
-  `${escapedRoot}\/packages\/(?!mobile|utils|walletkit|contractkit|react-components).*`
+  `${escapedRoot}\/packages\/(?!mobile|utils|contractkit|react-components).*`
 )
 const nestedRnRegex = new RegExp(`.*\/node_modules\/.*\/node_modules\/(react-native)\/.*`)
 const componentsRnRegex = new RegExp(`.*react-components\/node_modules\/(react-native)\/.*`)
