+++ packages/mobile/scripts/run_app.sh
@@ -70,13 +70,6 @@ rm -f $ENV_FILENAME.bak
 rm -f $ANDROID_GSERVICES_PATH.bak
 rm -f $IOS_GSERVICES_PATH.bak
 
-# Build Wallet Kit for env
-if [ "$FAST" = false ]; then
-  echo "Building sdk for testnet $NETWORK"
-  yarn build:sdk $NETWORK
-  echo "Done building sdk"
-fi
-
 # Build the app and run it
 if [ $PLATFORM = "android" ]; then
   echo "Using platform android"
