--- packages/walletkit/.gitignore
@@ -1,5 +1,0 @@
-.artifacts/
-contracts/
-dist/
-types/
-lib/
