--- packages/walletkit/package.json
@@ -1,57 +1,0 @@
-{
-  "name": "@celo/walletkit",
-  "version": "0.0.15",
-  "description": "Celo's WalletKit to interact with Celo network",
-  "main": "./lib/index.js",
-  "types": "./lib/index.d.ts",
-  "author": "Celo",
-  "license": "Apache-2.0",
-  "scripts": {
-    "build:for-env": "bin/build-sdk.js",
-    "build:ts": "tsc -p .",
-    "clean": "rm -rf lib .artifacts contracts types lib",
-    "lint": "tslint -c tslint.json --project tsconfig.json",
-    "prebuild": "yarn clean",
-    "build": "yarn run build:for-env alfajoresstaging",
-    "prepublishOnly": "yarn run build",
-    "test": "export TZ=UTC && jest --ci --silent --coverage --runInBand test/attestations.test.ts test/erc20-utils.test.ts test/google-storage-utils.test.ts test/start_geth.sh test/transaction-utils.test.ts",
-    "test:verbose": "export TZ=UTC && jest --ci --verbose --runInBand  test/attestations.test.ts test/erc20-utils.test.ts test/google-storage-utils.test.ts test/start_geth.sh test/transaction-utils.test.ts"
-  },
-  "bin": {
-    "build-sdk": "./bin/build-sdk.js"
-  },
-  "dependencies": {
-    "@0x/subproviders": "^4.1.0",
-    "@celo/utils": "0.1.9-dev",
-    "@google-cloud/storage": "^2.3.3",
-    "bignumber.js": "^7.2.0",
-    "chalk": "^2.4.2",
-    "cross-fetch": "^3.0.2",
-    "eth-lib": "^0.2.8",
-    "lodash": "^4.17.14",
-    "sleep-promise": "^8.0.1",
-    "web3": "1.2.4",
-    "web3-eth": "1.2.4",
-    "web3-utils": "1.2.4",
-    "web3-core": "1.2.4",
-    "web3-core-helpers": "1.2.4"
-  },
-  "devDependencies": {
-    "@types/bignumber.js": "^5.0.0",
-    "@types/bip32": "^1.0.1",
-    "@types/bip39": "^2.4.2",
-    "@types/lodash": "^4.14.136",
-    "bip32": "^1.0.2",
-    "bip39": "^2.5.0",
-    "chalk": "^2.4.2",
-    "dotenv": "6.1.0",
-    "typechain": "1.0.5",
-    "typechain-target-web3-v1": "1.0.4",
-    "typechain-target-truffle": "1.0.2"
-  },
-  "files": [
-    "index.ts",
-    "lib/**/*",
-    "bin/build-sdk.js"
-  ]
-}
