--- packages/walletkit/tsconfig.json
@@ -1,26 +1,0 @@
-{
-  "compileOnSave": false,
-  "compilerOptions": {
-    "allowSyntheticDefaultImports": true,
-    "baseUrl": ".",
-    "composite": true,
-    "declaration": true,
-    "jsx": "preserve",
-    "lib": ["dom", "es2015", "es2016"],
-    "module": "commonjs",
-    "moduleResolution": "node",
-    "noImplicitAny": true,
-    "noUnusedLocals": true,
-    "noUnusedParameters": true,
-    "outDir": "lib",
-    "removeComments": false,
-    "sourceMap": false,
-    "skipLibCheck": true,
-    "strict": true,
-    "target": "es5",
-    "esModuleInterop": true
-  },
-  "include": [".", "src/**/*", "types/**/*", "contracts/**/*"],
-  "exclude": ["lib", "node_modules"],
-  "references": [{ "path": "../utils" }]
-}
